39 Wilson Gardens
"That hurts my pride, Watson," he said at last. "It is a petty feeling, no doubt, but it hurts my
pride. It becomes a personal matter with me now, and, if God sends me health, I shall set my hand
upon this gang. That he should come to me for help, and that I should send him away to his death--!"
He sprang from his chair and paced about the room in uncontrollable agitation, with a flush upon his
sallow cheeks and a nervous clasping and unclasping of his long thin hands.

"They must be cunning devils," he exclaimed at last. "How could they have decoyed him down there?
The Embankment is not on the direct line to the station. The bridge, no doubt, was too crowded, even
on such a night, for their purpose. Well, Watson, we shall see who will win in the long run. I am
going out now!"

"To the police?"

"No; I shall be my own police. When I have spun the web they may take the flies, but not before."

All day I was engaged in my professional work, and it was late in the evening before I returned to
Baker Street. Sherlock Holmes had not come back yet. It was nearly ten o'clock before he entered,
looking pale and worn. He walked up to the sideboard, and tearing a piece from the loaf he devoured
it voraciously, washing it down with a long draught of water.