29 Glendale
"Oh, tut, tut! I have no time! That left foot of yours with its inward twist is all over the place.
A mole could trace it, and there it vanishes among the reeds. Oh, how simple it would all have been
had I been here before they came like a herd of buffalo and wallowed all over it. Here is where the
party with the lodge-keeper came, and they have covered all tracks for six or eight feet round the
body. But here are three separate tracks of the same feet." He drew out a lens and lay down upon his
waterproof to have a better view, talking all the time rather to himself than to us. "These are
young McCarthy's feet. Twice he was walking, and once he ran swiftly, so that the soles are deeply
marked and the heels hardly visible. That bears out his story. He ran when he saw his father on the
ground. Then here are the father's feet as he paced up and down. What is this, then? It is the butt-
end of the gun as the son stood listening. And this? Ha, ha! What have we here? Tiptoes! tiptoes!
Square, too, quite unusual boots! They come, they go, they come again--of course that was for the
cloak. Now where did they come from?" He ran up and down, sometimes losing, sometimes finding the
track until we were well within the edge of the wood and under the shadow of a great beech, the
largest tree in the neighbourhood. Holmes traced his way to the farther side of this and lay down
once more upon his face with a little cry of satisfaction. For a long time he remained there,
turning over the leaves and dried sticks, gathering up what seemed to me to be dust into an envelope
and examining with his lens not only the ground but even the bark of the tree as far as he could
reach. A jagged stone was lying among the moss, and this also he carefully examined and retained.
Then he followed a pathway through the wood until he came to the highroad, where all traces were
lost.

"It has been a case of considerable interest," he remarked, returning to his natural manner. "I
fancy that this grey house on the right must be the lodge. I think that I will go in and have a word
with Moran, and perhaps write a little note. Having done that, we may drive back to our luncheon.
You may walk to the cab, and I shall be with you presently."

It was about ten minutes before we regained our cab and drove back into Ross, Holmes still carrying
with him the stone which he had picked up in the wood.

"This may interest you, Lestrade," he remarked, holding it out.  "The murder was done with it."

"I see no marks."