5 Glan Road
"Of Friday, June 19th."

"Good heavens! I thought it was Wednesday. It is Wednesday. What d'you want to frighten a chap for?"
He sank his face onto his arms and began to sob in a high treble key.

"I tell you that it is Friday, man. Your wife has been waiting this two days for you. You should be
ashamed of yourself!"