50 Valentia Road
"I beg that you will draw your chair up to the fire and favour me with some details as to your
case."

"It is no ordinary one."

"None of those which come to me are. I am the last court of appeal."

"And yet I question, sir, whether, in all your experience, you have ever listened to a more
mysterious and inexplicable chain of events than those which have happened in my own family."

"You fill me with interest," said Holmes. "Pray give us the essential facts from the commencement,
and I can afterwards question you as to those details which seem to me to be most important."