8 Wantage Close
"No, I shall commence with the City. Just ring the bell and the maid will bring up your coffee."

As I waited, I lifted the unopened newspaper from the table and glanced my eye over it. It rested
upon a heading which sent a chill to my heart.

"Holmes," I cried, "you are too late."

"Ah!" said he, laying down his cup, "I feared as much. How was it done?" He spoke calmly, but I
could see that he was deeply moved.

"My eye caught the name of Openshaw, and the heading 'Tragedy Near Waterloo Bridge.' Here is the
account:

"Between nine and ten last night Police-Constable Cook, of the H Division, on duty near Waterloo
Bridge, heard a cry for help and a splash in the water. The night, however, was extremely dark and
stormy, so that, in spite of the help of several passers-by, it was quite impossible to effect a
rescue. The alarm, however, was given, and, by the aid of the water-police, the body was eventually
recovered. It proved to be that of a young gentleman whose name, as it appears from an envelope
which was found in his pocket, was John Openshaw, and whose residence is near Horsham. It is
conjectured that he may have been hurrying down to catch the last train from Waterloo Station, and
that in his haste and the extreme darkness he missed his path and walked over the edge of one of the
small landing-places for river steamboats. The body exhibited no traces of violence, and there can
be no doubt that the deceased had been the victim of an unfortunate accident, which should have the
effect of calling the attention of the authorities to the condition of the riverside landing-
stages."