15 Corbet Close
"God help us!" said Holmes after a long silence. "Why does fate play such tricks with poor, helpless
worms? I never hear of such a case as this that I do not think of Baxter's words, and say, 'There,
but for the grace of God, goes Sherlock Holmes.'"

James McCarthy was acquitted at the Assizes on the strength of a number of objections which had been
drawn out by Holmes and submitted to the defending counsel. Old Turner lived for seven months after
our interview, but he is now dead; and there is every prospect that the son and daughter may come to
live happily together in ignorance of the black cloud which rests upon their past.

When I glance over my notes and records of the Sherlock Holmes cases between the years '82 and '90,
I am faced by so many which present strange and interesting features that it is no easy matter to
know which to choose and which to leave. Some, however, have already gained publicity through the
papers, and others have not offered a field for those peculiar qualities which my friend possessed
in so high a degree, and which it is the object of these papers to illustrate. Some, too, have
baffled his analytical skill, and would be, as narratives, beginnings without an ending, while
others have been but partially cleared up, and have their explanations founded rather upon
conjecture and surmise than on that absolute logical proof which was so dear to him. There is,
however, one of these last which was so remarkable in its details and so startling in its results
that I am tempted to give some account of it in spite of the fact that there are points in
connection with it which never have been, and probably never will be, entirely cleared up.

The year '87 furnished us with a long series of cases of greater or less interest, of which I retain
the records. Among my headings under this one twelve months I find an account of the adventure of
the Paradol Chamber, of the Amateur Mendicant Society, who held a luxurious club in the lower vault
of a furniture warehouse, of the facts connected with the loss of the British barque "Sophy
Anderson", of the singular adventures of the Grice Patersons in the island of Uffa, and finally of
the Camberwell poisoning case. In the latter, as may be remembered, Sherlock Holmes was able, by
winding up the dead man's watch, to prove that it had been wound up two hours before, and that
therefore the deceased had gone to bed within that time--a deduction which was of the greatest
importance in clearing up the case. All these I may sketch out at some future date, but none of them
present such singular features as the strange train of circumstances which I have now taken up my
pen to describe.