50 Ventnor Terrace
"You fill me with interest," said Holmes. "Pray give us the essential facts from the commencement,
and I can afterwards question you as to those details which seem to me to be most important."

The young man pulled his chair up and pushed his wet feet out towards the blaze.

"My name," said he, "is John Openshaw, but my own affairs have, as far as I can understand, little
to do with this awful business. It is a hereditary matter; so in order to give you an idea of the
facts, I must go back to the commencement of the affair.