119 Valentia Road
I did so, and saw a large "E" with a small "g," a "P," and a large "G" with a small "t" woven into
the texture of the paper.

"What do you make of that?" asked Holmes.

"The name of the maker, no doubt; or his monogram, rather."

"Not at all. The 'G' with the small 't' stands for 'Gesellschaft,' which is the German for
'Company.' It is a customary contraction like our 'Co.' 'P,' of course, stands for 'Papier.' Now for
the 'Eg.' Let us glance at our Continental Gazetteer." He took down a heavy brown volume from his
shelves. "Eglow, Eglonitz--here we are, Egria. It is in a German-speaking country--in Bohemia, not
far from Carlsbad. 'Remarkable as being the scene of the death of Wallenstein, and for its numerous
glass-factories and paper-mills.' Ha, ha, my boy, what do you make of that?" His eyes sparkled, and
he sent up a great blue triumphant cloud from his cigarette.

"The paper was made in Bohemia," I said.