70 Moor End
"All right," said Jones with a stare and a snigger. "Well, would you please, sir, march upstairs,
where we can get a cab to carry your Highness to the police-station?"

"That is better," said John Clay serenely. He made a sweeping bow to the three of us and walked
quietly off in the custody of the detective.