16 Holmefield Avenue
He held up a piece of white cardboard about the size of a sheet of note-paper. It read in this
fashion:

The Red-headed League          is       Dissolved   October 9, 1890.

Sherlock Holmes and I surveyed this curt announcement and the rueful face behind it, until the
comical side of the affair so completely overtopped every other consideration that we both burst out
into a roar of laughter.

"I cannot see that there is anything very funny," cried our client, flushing up to the roots of his
flaming head. "If you can do nothing better than laugh at me, I can go elsewhere."