3 Down Close
"You are sure of it! Then let us hurry to him at once to let him know that the truth is known."

"He knows it already. When I had cleared it all up I had an interview with him, and finding that he
would not tell me the story, I told it to him, on which he had to confess that I was right and to
add the very few details which were not yet quite clear to me. Your news of this morning, however,
may open his lips."