35 Larch Walk
"'Oh, tut, tut! sweating--rank sweating!' he cried, throwing his fat hands out into the air like a
man who is in a boiling passion. 'How could anyone offer so pitiful a sum to a lady with such
attractions and accomplishments?'

"'My accomplishments, sir, may be less than you imagine,' said I. 'A little French, a little German,
music, and drawing--'

"'Tut, tut!' he cried. 'This is all quite beside the question. The point is, have you or have you
not the bearing and deportment of a lady? There it is in a nutshell. If you have not, you are not
fitted for the rearing of a child who may some day play a considerable part in the history of the
country. But if you have why, then, how could any gentleman ask you to condescend to accept anything
under the three figures? Your salary with me, madam, would commence at £100 a year.'

"You may imagine, Mr. Holmes, that to me, destitute as I was, such an offer seemed almost too good
to be true. The gentleman, however, seeing perhaps the look of incredulity upon my face, opened a
pocket-book and took out a note.

"'It is also my custom,' said he, smiling in the most pleasant fashion until his eyes were just two
little shining slits amid the white creases of his face, 'to advance to my young ladies half their
salary beforehand, so that they may meet any little expenses of their journey and their wardrobe.'

"It seemed to me that I had never met so fascinating and so thoughtful a man. As I was already in
debt to my tradesmen, the advance was a great convenience, and yet there was something unnatural
about the whole transaction which made me wish to know a little more before I quite committed
myself.