38 Wantage Close
"I do not know that there is anything else for me to explain. I was determined to preserve my
disguise as long as possible, and hence my preference for a dirty face. Knowing that my wife would
be terribly anxious, I slipped off my ring and confided it to the Lascar at a moment when no
constable was watching me, together with a hurried scrawl, telling her that she had no cause to
fear."

"That note only reached her yesterday," said Holmes.