4 Corbet Close
"'Death,' said he, and rising from the table he retired to his room, leaving me palpitating with
horror. I took up the envelope and saw scrawled in red ink upon the inner flap, just above the gum,
the letter K three times repeated. There was nothing else save the five dried pips. What could be
the reason of his overpowering terror? I left the breakfast-table, and as I ascended the stair I met
him coming down with an old rusty key, which must have belonged to the attic, in one hand, and a
small brass box, like a cashbox, in the other.

"'They may do what they like, but I'll checkmate them still,' said he with an oath. 'Tell Mary that
I shall want a fire in my room to-day, and send down to Fordham, the Horsham lawyer.'

"I did as he ordered, and when the lawyer arrived I was asked to step up to the room. The fire was
burning brightly, and in the grate there was a mass of black, fluffy ashes, as of burned paper,
while the brass box stood open and empty beside it. As I glanced at the box I noticed, with a start,
that upon the lid was printed the treble K which I had read in the morning upon the envelope.