1 Tamworth Drive
"Through the skylight. We shall soon see how he managed it." He swung himself up onto the roof. "Ah,
yes," he cried, "here's the end of a long light ladder against the eaves. That is how he did it."

"But it is impossible," said Miss Hunter; "the ladder was not there when the Rucastles went away."