11 Moor End
"There were none. I believe that it was an excuse to move me from my room."

"Ah! that is suggestive. Now, on the other side of this narrow wing runs the corridor from which
these three rooms open. There are windows in it, of course?"

"Yes, but very small ones. Too narrow for anyone to pass through."

"As you both locked your doors at night, your rooms were unapproachable from that side. Now, would
you have the kindness to go into your room and bar your shutters?"

Miss Stoner did so, and Holmes, after a careful examination through the open window, endeavoured in
every way to force the shutter open, but without success. There was no slit through which a knife
could be passed to raise the bar. Then with his lens he tested the hinges, but they were of solid
iron, built firmly into the massive masonry. "Hum!" said he, scratching his chin in some perplexity,
"my theory certainly presents some difficulties. No one could pass these shutters if they were
bolted. Well, we shall see if the inside throws any light upon the matter."

A small side door led into the whitewashed corridor from which the three bedrooms opened. Holmes
refused to examine the third chamber, so we passed at once to the second, that in which Miss Stoner
was now sleeping, and in which her sister had met with her fate. It was a homely little room, with a
low ceiling and a gaping fireplace, after the fashion of old country-houses. A brown chest of
drawers stood in one corner, a narrow white-counterpaned bed in another, and a dressing-table on the
left-hand side of the window. These articles, with two small wicker-work chairs, made up all the
furniture in the room save for a square of Wilton carpet in the centre. The boards round and the
panelling of the walls were of brown, worm-eaten oak, so old and discoloured that it may have dated
from the original building of the house. Holmes drew one of the chairs into a corner and sat silent,
while his eyes travelled round and round and up and down, taking in every detail of the apartment.