19 Cliff Place
"That was it, sir."

"But Mr. Fowler being a persevering man, as a good seaman should be, blockaded the house, and having
met you succeeded by certain arguments, metallic or otherwise, in convincing you that your interests
were the same as his."

"Mr. Fowler was a very kind-spoken, free-handed gentleman," said Mrs. Toller serenely.

"And in this way he managed that your good man should have no want of drink, and that a ladder
should be ready at the moment when your master had gone out."

"You have it, sir, just as it happened."

"I am sure we owe you an apology, Mrs. Toller," said Holmes, "for you have certainly cleared up
everything which puzzled us. And here comes the country surgeon and Mrs. Rucastle, so I think,
Watson, that we had best escort Miss Hunter back to Winchester, as it seems to me that our locus
standi now is rather a questionable one."