3 Glan Road
"I think that this should do," said he, glancing into the glass above the fireplace. "I only wish
that you could come with me, Watson, but I fear that it won't do. I may be on the trail in this
matter, or I may be following a will-o'-the-wisp, but I shall soon know which it is. I hope that I
may be back in a few hours." He cut a slice of beef from the joint upon the sideboard, sandwiched it
between two rounds of bread, and thrusting this rude meal into his pocket he started off upon his
expedition.

I had just finished my tea when he returned, evidently in excellent spirits, swinging an old
elastic-sided boot in his hand. He chucked it down into a corner and helped himself to a cup of tea.

"I only looked in as I passed," said he. "I am going right on."

"Where to?"

"Oh, to the other side of the West End. It may be some time before I get back. Don't wait up for me
in case I should be late."

"How are you getting on?"