25 Harvey Street
"I cannot imagine."

"No, I don't think you would guess. Every pocket stuffed with pennies and half-pennies--421 pennies
and 270 half-pennies. It was no wonder that it had not been swept away by the tide. But a human body
is a different matter. There is a fierce eddy between the wharf and the house. It seemed likely
enough that the weighted coat had remained when the stripped body had been sucked away into the
river."

"But I understand that all the other clothes were found in the room. Would the body be dressed in a
coat alone?"