68 Corbet Close
"I have sworn it by the most solemn oaths which a man can take."

"In that case I think that it is probable that no further steps may be taken. But if you are found
again, then all must come out. I am sure, Mr. Holmes, that we are very much indebted to you for
having cleared the matter up. I wish I knew how you reach your results."

"I reached this one," said my friend, "by sitting upon five pillows and consuming an ounce of shag.
I think, Watson, that if we drive to Baker Street we shall just be in time for breakfast."