59 Down Close
"'What are you doing there?' he asked.

"I felt angry at having been tricked by so elaborate a story as that which he had told me. 'I was
admiring your fuller's-earth,' said I; 'I think that I should be better able to advise you as to
your machine if I knew what the exact purpose was for which it was used.'