4 Tamworth Drive
"Dearest do not be frightened. All will come well. There is a huge error which it may take some
little time to rectify. Wait in patience. "Neville.

Written in pencil upon the fly-leaf of a book, octavo size, no water-mark. Hum! Posted to-day in
Gravesend by a man with a dirty thumb. Ha! And the flap has been gummed, if I am not very much in
error, by a person who had been chewing tobacco. And you have no doubt that it is your husband's
hand, madam?"