147 Bowler Street
"The name, you see, is in perfectly black ink, which has dried itself. The rest is of the greyish
colour, which shows that blotting-paper has been used. If it had been written straight off, and then
blotted, none would be of a deep black shade. This man has written the name, and there has then been
a pause before he wrote the address, which can only mean that he was not familiar with it. It is, of
course, a trifle, but there is nothing so important as trifles. Let us now see the letter. Ha!
there has been an enclosure here!"

"Yes, there was a ring. His signet-ring."

"And you are sure that this is your husband's hand?"

"One of his hands."

"One?"

"His hand when he wrote hurriedly. It is very unlike his usual writing, and yet I know it well."